[54] Fix of Most expensive category bug
