+++ ui/src/main/java/br/com/tick/ui/screens/analysis/states/MostExpensiveCategoriesStates.kt
@@ -7,7 +7,17 @@ sealed class MostExpensiveCategoriesStates {
     companion object {
         fun of(mostExpensiveCategories: List<MostExpensiveCategory>): MostExpensiveCategoriesStates {
 
-            return if (mostExpensiveCategories.isEmpty()) Loading else Full(mostExpensiveCategories.take(5))
+            return if (mostExpensiveCategories.isEmpty()) {
+                Loading
+            } else {
+                Full(
+                    mostExpensiveCategories
+                        .sortedByDescending {
+                            it.amount
+                        }
+                        .take(5)
+                )
+            }
         }
     }
